+++ src/main/java/io/github/sakurawald/module/initializer/world/structure/DimensionNode.java
@@ -1,6 +1,8 @@
 package io.github.sakurawald.module.initializer.world.structure;
 
 import com.google.gson.annotations.SerializedName;
+import io.github.sakurawald.core.auxiliary.minecraft.RegistryHelper;
+import io.github.sakurawald.core.auxiliary.minecraft.ServerHelper;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -13,4 +15,8 @@ public class DimensionNode {
     @SerializedName(value = "dimension_type", alternate = "dimensionType")
     String dimension_type;
     long seed;
+
+    public boolean isDimensionLoaded() {
+        return ServerHelper.getWorlds().stream().anyMatch(it -> RegistryHelper.ofString(it).equals(this.dimension));
+    }
 }