+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 > The version number of fuji follows `semver` now: https://semver.org/ 
  
 - i18n: new translation for `zh_tw`. @yichifauzi
+- feature: add `gui support` for `/world list`. (world module)
+- fix: the `next page button` not work in `/works` gui. (works module).